Breakup functions and utils This will allow individual functions to be imported for upcoming import functionality
