+++ src/helpers/constants.js
@@ -1,0 +1,29 @@
+module.exports = {
+    NATIVE_FUNCTIONS: {
+        'decodeURI': 1,
+        'decodeURIComponent': 1,
+        'encodeURI': 1,
+        'encodeURIComponent': 1,
+        'isFinite': 1,
+        'isNaN': 1,
+        'parseInt': 1,
+        'parseFloat': 1
+    },
+
+    NATIVE_OBJECTS: {
+        'Array': 1,
+        'Boolean': 1,
+        'Date': 1,
+        'Error': 1,
+        'Function': 1,
+        'JSON': 1,
+        'Math': 1,
+        'Map': 1,
+        'Number': 1,
+        'Object': 1,
+        'RegExp': 1,
+        'String': 1,
+        'WeakMap': 1,
+        'console': 1
+    }
+}
